Move `format_table` to `util.helper`
